refactor: SignUpStepScaffold 파일 위치 이동 및 패키지 구조 정리
- `SignUpStepScaffold.kt`를 `component` 하위 패키지에서 `signup` 패키지로 이동
- 파일 이동에 따른 `SignUpPasswordScreen`, `SignUpResidentNumberScreen`, `SignUpScreen`의 임포트 경로 수정 및 불필요한 임포트 제거
- `SignUpStepScaffold` 내 불필요한 `SIGN_UP_TOTAL_STEPS` 임포트 제거
